Backend API restructured
The backend API has been completely restructured to include seperate models
for the following fields:

* Researchers
* Isotopes
* Parameters

This will make filtering faster in the future and will make it easier to add
additional attributes. In addition, changes have been made to the admin page
so that it is easier to create documents. This is a feature exclusive to
admins. Lastly, changes have been made to use the ListAPIView, this
makes pagination and filtering simpler to implement.