Implemented frontend http requests to consume backend API
Frontend is now capable of consuming the backend API using 'axios'. Pagaination
and page loading is working correctly. However, filtering still needs to
be implemented. In addition, changes to the frontend have been made to include
fontawesome icons, deleted unnecessary assets and a central theme which can be
adjusted using the .scss file.